# ARROW-9949: [C++] Improve performance of Decimal128::FromString by 46%, and make the implementation reusable for Decimal256.