```
Running release/arrow-decimal-benchmark
Run on (12 X 4500 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 1024 KiB (x6)
  L3 Unified 8448 KiB (x1)
Load Average: 0.33, 1.08, 0.93
---------------------------------------------------------------------
Benchmark                    Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------
FromString (before)        263 ns          263 ns      2666389 items_per_second=22.8378M/s
FromString (after)         180 ns          180 ns      3873651 items_per_second=33.3283M/s
```