@pitrou 2020-09-09
Be careful: it seems you updated the `testing` submodule.