@pitrou 2020-09-10
For the record, I get a ~20% speedup on AMD Ryzen.