@MingyuZhong 2020-09-10
> Ouch, this accidentally broke the ASAN/UBSAN build and I hadn't seen it because of other (unrelated) CI failures:
> https://github.com/apache/arrow/runs/1093237142
> 
> ```
> src/arrow/util/CMakeFiles/arrow-utility-test.dir/decimal_test.cc.o: In function `arrow::Decimal128Test_Multiply_Test::TestBody()':
> /arrow/cpp/src/arrow/util/decimal_test.cc:935: undefined reference to `__muloti4'
> clang: error: linker command failed with exit code 1 (use -v to see invocation)
> ```

I think we hit https://stackoverflow.com/questions/49793632/clang-fsanitize-undefined-with-128-integer-operations-undefined-reference-to . Is there a way to use compiler-rt for ASAN/UBSAN? If not, we can remove that line or revert to always use boost.