@pitrou 2020-09-10
I'll post a PR for a quick fix.